@@ -12,26 +12,26 @@
 		"format": "prettier --plugin-search-dir . --write ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.2.5",
-		"@sveltejs/adapter-static": "^3.0.5",
-		"@sveltejs/kit": "^2.6.4",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"@typescript-eslint/eslint-plugin": "^8.8.1",
-		"@typescript-eslint/parser": "^8.8.1",
+      "@sveltejs/adapter-auto": "^4.0.0",
+      "@sveltejs/adapter-static": "^3.0.8",
+      "@sveltejs/kit": "^2.17.1",
+      "@sveltejs/vite-plugin-svelte": "^5.0.3",
+      "@typescript-eslint/eslint-plugin": "^8.23.0",
+      "@typescript-eslint/parser": "^8.23.0",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.12.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.44.1",
-		"postcss": "^8.4.47",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.7",
-		"prettier-plugin-tailwindcss": "^0.6.8",
-		"svelte": "^4.2.19",
-		"svelte-check": "^4.0.4",
-		"tailwindcss": "^3.4.13",
-		"tslib": "^2.7.0",
-		"typescript": "^5.6.3",
-		"vite": "^5.4.8"
+      "eslint": "^9.19.0",
+      "eslint-config-prettier": "^10.0.1",
+      "eslint-plugin-svelte": "^2.46.1",
+      "postcss": "^8.5.1",
+      "prettier": "^3.4.2",
+      "prettier-plugin-svelte": "^3.3.3",
+      "prettier-plugin-tailwindcss": "^0.6.11",
+      "svelte": "^5.19.9",
+      "svelte-check": "^4.1.4",
+      "tailwindcss": "^4.0.4",
+      "tslib": "^2.8.1",
+      "typescript": "^5.7.3",
+      "vite": "^6.1.0"
 	},
 	"type": "module"
 }
