@@ -10,14 +10,14 @@
 	import Soundcloud from '$components/soundcloud.svelte';
 	import { howToPlay_entries } from '$lib/howToPlayEntries';
 
-	let guesses: SongContainerArray = [
+	let guesses: SongContainerArray = $state([
 		'nuetral',
 		'nuetral',
 		'nuetral',
 		'nuetral',
 		'nuetral',
 		'nuetral'
-	];
+	]);
 
 	let url = 'https://soundcloud.com/black-eyed-peas-official/meet-me-halfway-album-version';
 </script>
@@ -27,39 +27,46 @@
 
 	{#if $showHTPModal}
 		<Modal title={'HOW TO PLAY'}>
-			<div slot="body">
-				{#each howToPlay_entries as { svg, text }, index}
-					<div class="mb-6 flex items-center">
-						<img src={svg} alt="" class="mr-4 size-7" />
-						<div class="tracking-wide">
-							{text}
+			{#snippet body()}
+				<div>
+					{#each howToPlay_entries as { svg, text }, index}
+						<div class="mb-6 flex items-center">
+							<img src={svg} alt="" class="mr-4 size-7" />
+							<div class="tracking-wide">
+								{text}
+							</div>
 						</div>
-					</div>
-				{/each}
-			</div>
-			<button
-				slot="footer"
-				class="bg-green-500 px-2 py-[10px] text-sm font-semibold tracking-wider text-normal"
-				on:click={() => showHTPModal.set(false)}
-			>
-				PLAY
-			</button>
+					{/each}
+				</div>
+			{/snippet}
+			{#snippet footer()}
+				<button
+
+					class="bg-green-500 px-2 py-[10px] text-sm font-semibold tracking-wider text-normal"
+					onclick={() => showHTPModal.set(false)}
+				>
+					PLAY
+				</button>
+			{/snippet}
 		</Modal>
 	{/if}
 	{#if $showAboutModal}
 		<Modal title="About">
-			<div slot="body" class="text-center font-noto text-[0.95rem]">
-				Each daily Heardle features a clip from a popular song. Guess in as few tries as possible,
-				and be sure to come back every day for a new song.
-				<br />
-				<br />
-				Based on Heardle.app which was unfortunately shut down.
-				<br />
-				<br />
-			</div></Modal
+			{#snippet body()}
+				<div class="text-center font-noto text-[0.95rem]">
+					Each daily Heardle features a clip from a popular song. Guess in as few tries as possible,
+					and be sure to come back every day for a new song.
+					<br />
+					<br />
+					Based on Heardle.app which was unfortunately shut down.
+					<br />
+					<br />
+				</div>
+			{/snippet}
+		</Modal
 		>
 	{/if}
-	<SongEntry bind:guesses />
+	<SongEntry {guesses} />
 	<ProgressBar />
 	<Soundcloud
 		hidden="true"
